move card fetching to card.elm
because a keypress from the input triggers grabbing the card I initially
had the card fetching logic in input.elm it makes more sense to have
this logic in card.elm. I still have to resolve some of the strangeness
around how actions are passed to parents and across sub components  #
Please enter the commit message for your changes. Lines starting # with
'#' will be ignored, and an empty message aborts the commit.  # On
branch master # Your branch is up-to-date with 'origin/master'.  # #
Changes to be committed: #	modified:   Card.elm #	modified:
Input.elm #	modified:   Main.elm #